chore: rename component files
