+++ frontend/packages/design-system/src/components/gantt-view/gantt-bar/projectBar.tsx
@@ -1,6 +1,6 @@
-import { useGanttStore } from "../gantt-store";
+import { useGanttStore } from "../ganttStore";
 import type { Allocation } from "../types";
-import { GanttBar } from "./gantt-bar";
+import { GanttBar } from "./ganttBar";
 import { getClampedBarLayout } from "./utilities/getClampedBarLayout";
 import { getNumDays } from "./utilities/getNumDays";
