+++ frontend/packages/design-system/src/components/gantt-view/ganttProjectItem.tsx
