+++ frontend/packages/design-system/src/components/gantt-view/ganttStore.ts
