+++ frontend/packages/design-system/src/components/gantt-view/index.ts
@@ -1,2 +1,2 @@
-export * from "./gantt-grid";
+export * from "./ganttGrid";
 export * from "./types";
